chore: remover fixtures que nao eram utilizadas no teste da chamada no use
criar FixtureCenarioBaseTestCase para organizar metodos que utilizam fixture

remover codigo que estava comentado
