+++ tests_sei4/funcional/bootstrap.php
@@ -39,3 +39,4 @@ require_once __DIR__ . '/src/paginas/PaginaCadastroMapEnvioCompDigitais.php';
 require_once __DIR__ . '/src/paginas/PaginaEnvioParcialListar.php';
 
 require_once __DIR__ . '/tests/CenarioBaseTestCase.php';
+require_once __DIR__ . '/tests/FixtureCenarioBaseTestCase.php';
