+++ tests_sei4/funcional/phpunit.xml
@@ -36,17 +36,17 @@
     <const name="CONTEXTO_ORGAO_A_REP_ESTRUTURAS" value="RE CGPRO" />
     <const name="CONTEXTO_ORGAO_A_SIGLA_UNIDADE" value="TESTE" />
     <const name="CONTEXTO_ORGAO_A_ID_ESTRUTURA"
-      value="[INFORME O ID DE ESTRUTURA UTILIZADO PARA TESTE ORG1]" />
+      value="157402" />
     <const name="CONTEXTO_ORGAO_A_SIGLA_ESTRUTURA"
-      value="[INFORME A SIGLA DE ESTRUTURA UTILIZADO PARA TESTE ORG1]" />
+      value="MARCUS_FIRST_ORG_1" />
     <const name="CONTEXTO_ORGAO_A_SIGLA_UNIDADE_HIERARQUIA" value="" />
     <const name="CONTEXTO_ORGAO_A_NOME_UNIDADE"
-      value="[INFORME O NOME DA ESTRUTURA UTILIZADO PARA TESTE ORG1]" />
+      value="MARCUS_FIRST_ORG_1" />
     <const name="CONTEXTO_ORGAO_A_SIGLA_UNIDADE_SECUNDARIA" value="TESTE_1_1" />
     <const name="CONTEXTO_ORGAO_A_ID_ESTRUTURA_SECUNDARIA"
-      value="[INFORME O ID DE ESTRUTURA UTILIZADO PARA TESTE ORG 1.11]" />
+      value="157405" />
     <const name="CONTEXTO_ORGAO_A_NOME_UNIDADE_SECUNDARIA"
-      value="[INFORME O NOME DA ESTRUTURA UTILIZADO PARA TESTE ORG 1.1]" />
+      value="MARCUS_FIRST_ORG_1_F" />
     <const name="CONTEXTO_ORGAO_A_SIGLA_UNIDADE_SECUNDARIA_HIERARQUIA" value="" />
     <const name="CONTEXTO_ORGAO_A_USUARIO_LOGIN" value="teste" />
     <const name="CONTEXTO_ORGAO_A_USUARIO_SENHA" value="teste" />
@@ -78,12 +78,12 @@
     <const name="CONTEXTO_ORGAO_B_REP_ESTRUTURAS" value="RE CGPRO" />
     <const name="CONTEXTO_ORGAO_B_SIGLA_UNIDADE" value="TESTE" />
     <const name="CONTEXTO_ORGAO_B_ID_ESTRUTURA"
-      value="[INFORME O ID DE ESTRUTURA UTILIZADO PARA TESTE ORG 2]" />
+      value="157403" />
     <const name="CONTEXTO_ORGAO_B_SIGLA_ESTRUTURA"
-      value="[INFORME A SIGLA DE ESTRUTURA UTILIZADO PARA TESTE ORG2]" />
+      value="MARCUS_FIRST_ORG_2" />
     <const name="CONTEXTO_ORGAO_B_SIGLA_UNIDADE_HIERARQUIA" value="" />
     <const name="CONTEXTO_ORGAO_B_NOME_UNIDADE"
-      value="[INFORME O NOME DA ESTRUTURA UTILIZADO PARA TESTE ORG 2]" />
+      value="MARCUS_FIRST_ORG_2" />
     <const name="CONTEXTO_ORGAO_B_USUARIO_LOGIN" value="teste" />
     <const name="CONTEXTO_ORGAO_B_USUARIO_SENHA" value="teste" />
     <const name="CONTEXTO_ORGAO_B_ID_ESTRUTURA_SECUNDARIA" value="XXXXXXXXXXXXXX" />
