+++ tests_super/funcional/tests/TramiteProcessoComDevolucaoAnexadoOutroTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use Tests\Funcional\Sei\Fixtures\{ProtocoloFixture,ProcedimentoFixture,AtividadeFixture,ContatoFixture,ParticipanteFixture,RelProtocoloAssuntoFixture,AtributoAndamentoFixture,DocumentoFixture,AssinaturaFixture,AnexoFixture,AnexoProcessoFixture};
+use Tests\Funcional\Sei\Fixtures\ProtocoloFixture;
 
 /**
  * Testes de trâmite de um processo tendo a sua devolução através de sua anexação à outro processo
@@ -12,7 +12,7 @@ use Tests\Funcional\Sei\Fixtures\{ProtocoloFixture,ProcedimentoFixture,Atividade
  * Execution Groups
  * @group execute_alone_group5
  */
-class TramiteProcessoComDevolucaoAnexadoOutroTest extends CenarioBaseTestCase
+class TramiteProcessoComDevolucaoAnexadoOutroTest extends FixtureCenarioBaseTestCase
 {
     public static $remetente;
     public static $destinatario;
@@ -107,8 +107,6 @@ class TramiteProcessoComDevolucaoAnexadoOutroTest extends CenarioBaseTestCase
 
         // Cadastra processo principal, seus documentos e anexa processo recebido anteriormente
         $objProtocoloPrincipalDTO = $this->cadastrarProcessoFixture(self::$processoTestePrincipal);
-        // $objProtocoloPrincipalDTO = $DTOs['Protocolo'];
-        // $objProcedimentoPrincipalDTO = $DTOs['Procedimento'];
         self::$protocoloTestePrincipal = $objProtocoloPrincipalDTO->getStrProtocoloFormatado(); 
 
         // Cadastra e assina
@@ -118,7 +116,6 @@ class TramiteProcessoComDevolucaoAnexadoOutroTest extends CenarioBaseTestCase
 
         $this->anexarProcessoFixture($objProtocoloPrincipalDTO->getDblIdProtocolo(), $objProtocoloAnexadoDTO->getDblIdProtocolo());
 
-        // $this->cadastrarDocumentoExternoFixture(self::$documentoTeste5);
         $this->cadastrarDocumentoExternoFixture(self::$documentoTeste5, $objProtocoloPrincipalDTO->getDblIdProtocolo());
         
         putenv("DATABASE_HOST=org1-database");
